feat: JPA paging변경 및 service 형태를 interface로 변경
