+++ src/main/java/com/ohgiraffers/warehousemanagement/wms/inspection/repository/InspectionRepository.java
@@ -2,13 +2,15 @@ package com.ohgiraffers.warehousemanagement.wms.inspection.repository;
 
 import com.ohgiraffers.warehousemanagement.wms.inspection.model.entity.Inspection;
 import com.ohgiraffers.warehousemanagement.wms.inspection.model.common.InspectionTransactionType;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
 import java.util.Optional;
 
 public interface InspectionRepository extends JpaRepository<Inspection, Integer> {
+
     Optional<Inspection> findByTransactionTypeAndTransactionId(InspectionTransactionType transactionType, Integer transactionId);
 
-    List<Inspection> findAllByTransactionType(InspectionTransactionType transactionType);
+    Page<Inspection> findAllByTransactionType(InspectionTransactionType transactionType, Pageable pageable);
 }
